Move AmplitudeSnippet to amplitude package.
