+++ project/settings.py
@@ -185,7 +185,7 @@ TEMPLATES = [
                 "project.context_processors.gtm_noscript_snippet",
                 "project.context_processors.facebook_pixel_snippet",
                 "project.context_processors.facebook_pixel_noscript_snippet",
-                "project.context_processors.amplitude_snippet",
+                "amplitude.context_processors.amplitude_snippet",
                 "project.context_processors.fullstory_snippet",
                 "project.context_processors.rollbar_snippet",
             ],
